+++ src/routes/[[lang]]/(auth)/login/_lib/index.ts
@@ -6,12 +6,13 @@ import { pocketbase } from '@slangroom/pocketbase';
 import { writable } from 'svelte/store';
 import scriptGenerateUser from './scriptGenerateUser.zen?raw';
 import scriptGenerateDid from './scriptGenerateDid.zen?raw';
+import { setUser } from '$lib/preferences/user';
 
 //
 
 const slangroom = new Slangroom(pocketbase);
 
-const pb_address: string = 'https://admin.signroom.io';
+const pb_address: string = 'https://admin.didroom.com';
 const password = 'CiccioLiam12345!'
 
 export const userEmailStore = writable<{email:string | undefined, registration:boolean}>();
@@ -55,19 +56,44 @@ export const generateDid = async (email:string) => {
 		url: '/api/did',
 	};
 
+	type User = {
+		avatar: string
+		bitcoin_public_key: string;
+		collectionId: string;
+		collectionName: string;
+		created: string;
+		ecdh_public_key: string;
+		eddsa_public_key: string;
+		email: string;
+		emailVisibility: boolean;
+		es256_public_key: string;
+		ethereum_address: string;
+		id: string;
+		name: string;
+		reflow_public_key: string;
+		updated: string;
+		username: string;
+		verified: boolean;
+	};
+
 	type DIDResponse = {
 		result: {
 			output: {
 				created: boolean;
 				did: object;
 			};
+			login_output: {
+				record: User;
+			}
 		};
 	};
 
 	const res = (await slangroom.execute(scriptGenerateDid, {
 		data
 	})) as unknown as DIDResponse;
-
+	
+	const loginOutput = res.result.login_output.record;
+	await setUser(loginOutput.id, email, loginOutput.name, loginOutput.avatar);
 	await setDIDPreference(res.result.output.did);
 
 	return res.result.output;