+++ src/routes/[[lang]]/(protected)/[id]/credential-detail/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import Header from '$lib/components/molecules/Header.svelte';
-	import { r } from '$lib/i18n';
+	import ScanButton from '$lib/components/molecules/ScanButton.svelte';
 	export let data: any;
 	const { credential, credentials } = data;
 	import { register } from 'swiper/element/bundle';
@@ -31,13 +31,12 @@
 		class="mt-8"
 	>
 		{#each credentials as credential}
-			<swiper-slide id={getIndex(credential.id)}
-				><d-credential-card
+			<swiper-slide id={getIndex(credential.id)}>
+        <d-credential-card
 					{...credential}
+					name={credential.display_name}
+			    logoSrc={credential.logo.url}
 					description=""
-					issuer={credential.issuer.length > 18
-						? credential.issuer.slice(0, 18) + '...'
-						: credential.issuer}
 				/>
 			</swiper-slide>
 		{/each}
@@ -46,9 +45,10 @@
 		<d-credential-detail
 			{...detailCredential}
 			description={detailCredential.description}
+			name={detailCredential.display_name}
 			logoSrc={detailCredential.logo.url}
 		>
-			<d-button color="accent" href={r('/scan/')} expand>SCAN FOR VERIFICATION</d-button>
 		</d-credential-detail>
 	</div>
+	<ScanButton />
 </ion-content>