+++ src/routes/[[lang]]/(protected)/credential-offer-test/+page.svelte
@@ -52,7 +52,7 @@
 		setTimeout(async () => {
 			const savedCredential = await setCredentialPreference({
 				configuration_ids: parsedService.credential_configuration_ids,
-				name: qrToWellKnown.credential_requested.display[0].name,
+				display_name: qrToWellKnown.credential_requested.display[0].name,
 				sdJwt: serviceResponse.credential,
 				issuer: parsedService.credential_issuer,
 				description: qrToWellKnown.credential_requested.display[0].description,
@@ -137,9 +137,11 @@
 							<div class="flex w-full justify-around">
 								<ion-icon icon={thumbsUpOutline} class="mx-auto my-6 text-9xl text-green-400"
 								></ion-icon>
-								<pre>
-								{JSON.stringify(serviceResponse, null, 2)}
-							</pre>
+								<div class="flex flex-col gap-2 ion-padding">
+									<d-text class="break-words">credential: {serviceResponse.credential}</d-text>
+									<d-text class="break-words">c_nonce: {serviceResponse.c_nonce}</d-text>
+									<d-text class="break-words">c_nonce_expires_in: {serviceResponse.c_nonce_expires_in}</d-text>
+							</div>
 							</div>
 						{/if}
 					</div>